fix sintax

+++ ft_substr.c
@@ -9,9 +9,9 @@ char *ft_substr(char const *s, unsigned int start, size_t len)
 
     if (!s)
         return (NULL);
-    if (start > ft_strlen(s))
+    if ((int)start > ft_strlen(s))
         return (ft_strdup(""));
-    if (len > ft_strlen(s + start))
+    if ((int)len > ft_strlen(s + start))
         len = ft_strlen(s + start);
     str = (char *)malloc(sizeof(*str) * (len + 1));
     if (!str)
